Move model methods into mixin
